agent.launch: add a robot_state_publisher for each agent to publish many of the gazebo model's transforms
Also adds static transforms between chassis and the stationary link of the four shock absorbers

I tried accomplishing this in the urdf, but (left|right)_steering_link and (left|right)_rear_axle_carrier refused to connect to the agent's chassis frame.

This is fix is less than ideal but has the desired effect, allowing the RobotModel to be viewed in rviz for simulated hamsters.